Adding support for a bulk downloader based on a CSV upload.

+++ src/main/java/org/barcodeapi/server/cache/CachedBarcode.java
@@ -26,7 +26,7 @@ public class CachedBarcode extends CachedObject {
 		return cachedData;
 	}
 
-	public long getDataSize() {
+	public int getDataSize() {
 
 		return cachedData == null ? 0 : cachedData.length;
 	}
